REPL: specify regions with open-close ranges using Pairs
All call sites of `splice_buffer!` were manually substracting 1 to
the last position of the range, indicating that it's more natural
to specify a region with an open-close range.
Therefore, `splice_buffer!` and `edit_replace` (thin wrapper of
`splice_buffer!`) are merged into `edit_splice!`, which takes
a `Region` (a pair of integers) object instead of a range.